Produce DLL files when compiled for windows
When we cross-compile telepathy-glib for windows with the current
flags we end up with a static library and it creates link issues
in libraries and applications which uses teleapthy-glib on windows.
to solve this -no-undefined LDFLAG is introduced in this commit.
This ensures all the symbols are defined when we compile for windows.

see: http://www.gnu.org/software/libtool/manual/libtool.html for
more details about -no-undefined flag.